Move CitationRow to its own component
